Add NPM support
This should (theoretically) allow `require(‘alertifyjs’)` to work when
someone using browserify installs alertifyjs via npm.

.npmignore is mostly a copy of .gitignore but also ignores html
documentation sources which is not needed in the package.

src/ is not ignored (even though src/js/ is not in a format that can be
used), in case someone wishes to import the LESS files.